Reorganize channel selection range code
